+++ src/my_neural_network/neuron.py
@@ -36,10 +36,10 @@ class SimpleNeuralNetwork:
         np.random.seed(self.config.seed)
         for l in range(1, self.L):
             self.parameters["W" + str(l)] = (
-                np.random.randn(self.layer_dims[l], self.layer_dims[l - 1]) * 0.01  # * np.sqrt(2 / self.layer_dims[l - 1]) 
+                np.random.randn(self.layer_dims[l], self.layer_dims[l - 1])
+                * 0.01  # * np.sqrt(2 / self.layer_dims[l - 1])
             )
             self.parameters["b" + str(l)] = np.zeros((self.layer_dims[l], 1))
-
 
     def _initialize_adam(self):
         """
@@ -52,7 +52,6 @@ class SimpleNeuralNetwork:
             self.m["db" + str(l)] = np.zeros_like(self.parameters["b" + str(l)])
             self.v["dW" + str(l)] = np.zeros_like(self.parameters["W" + str(l)])
             self.v["db" + str(l)] = np.zeros_like(self.parameters["b" + str(l)])
-
 
     def compute_loss(self, AL: np.ndarray, Y: np.ndarray) -> float:
         """
@@ -78,7 +77,6 @@ class SimpleNeuralNetwork:
         else:  # categorical cross-entropy for multi-class
             cost = -np.sum(Y * np.log(AL + 1e-8)) / m
         return np.squeeze(cost)  # reduce the dimensionality of the cost to a scalar
-
 
     def forward_propagation(self, X: np.ndarray) -> tuple[np.ndarray, list]:
         """
@@ -158,7 +156,6 @@ class SimpleNeuralNetwork:
         else:  # Multi-class classification- softmax activation
             return AL - Y
 
-
     def backward_propagation(self, AL: np.ndarray, Y: np.ndarray, caches: list) -> dict:
         """
         Perform the backward pass using gradient descent to compute the gradients of the loss function
@@ -212,7 +209,6 @@ class SimpleNeuralNetwork:
 
         return grads
 
-
     def update_parameters(self, grads: dict, t: Optional[int]) -> None:
         """
         Update the parameters using gradient descent.
@@ -231,7 +227,6 @@ class SimpleNeuralNetwork:
                 self.parameters["b" + str(l)] -= learning_rate * grads["db" + str(l)]
         elif optimizer == "adam":
             self.parameters = self._update_parameters_with_adam(t)
-
 
     def _update_parameters_with_adam(self, grads, t):
         """
@@ -279,7 +274,6 @@ class SimpleNeuralNetwork:
                 m_corrected_db / (np.sqrt(v_corrected_db) + epsilon)
             )
 
-
     def train(self, X: np.ndarray, Y: np.ndarray, epochs: int) -> None:
         """
         This method trains the neural network using gradient descent optimization.
@@ -320,7 +314,6 @@ class SimpleNeuralNetwork:
 
             # if i % 100 == 0:  # Print the cost every 100 epochs
             #     print(f"Cost after iteration {i}: {cost}")
-
 
     def predict(self, X: np.ndarray) -> np.ndarray:
         """